Added a new .po header called "Translators"
This modification to the stable version of Pootle 2.5.0 adds some more
appropriate header data to output .po files.  Instead of only listing
the last translator, it adds a new header called "Translators".  This
produces a semi-colon delimited list of all users who have submitted
translations for the corresponding language, in all projects.  It also
modifies the "sync_stores" administration command, to also ensure the
headers of a .po file are up to date, rather than just the translations.

The goal of this commit was to achieve an automated credits menu for the
Multi Theft Auto (MTA) project (http://code.google.com/p/mtasa-blue/).
When a user access the 'About' menu, a section of credits under
"Language localization" appears which is produced at runtime based upon
the header of the .po file.  This way, all users are credited without
constantly having to update the .po file or the source code of the
original software.

There are some very MTA-centric design aspects to this modification,
which should probably be modified.  The first one is the format of the
translator credit, which is '[First Name] "[Username]" [Last Name]'.
This matches how credits are structured in the MTA about screen.

The other MTA-centric design is the fact that this produces a list of
submissions for a specific language under /all/ projects, rather than
the translation project the .po file corresponds to.  This is due to how
the MTA Pootle integration works, but is probably not practical for most
uses.   This can be remedied with small modifications to the changes in
top_stats.py by using a pootleprofile__submission__translation_project
filter rather than
pootleprofile__submission__translation_project__language filter.